Marshal error messages across OOP as well
